Move Connectors base classes into diagram package
