+++ gaphor/adapters/__init__.py
@@ -1,4 +1,3 @@
-import gaphor.adapters.connectors
 import gaphor.adapters.editors
 import gaphor.adapters.grouping
