+++ gaphor/diagram/general/tests/test_comment.py
@@ -3,9 +3,12 @@ Comment and comment line items connection adapters tests.
 """
 
 from gaphor import UML
-from gaphor.diagram import items
-
 from gaphor.tests import TestCase
+from ..comment import CommentItem
+from ..commentline import CommentLineItem
+from ...usecases.actor import ActorItem
+from ...classes.klass import ClassItem
+from ...classes.generalization import GeneralizationItem
 
 
 class CommentLineTestCase(TestCase):
@@ -19,8 +22,8 @@ class CommentLineTestCase(TestCase):
     def test_commentline_annotated_element(self):
         """Test comment line item annotated element creation
         """
-        comment = self.create(items.CommentItem, UML.Comment)
-        line = self.create(items.CommentLineItem)
+        comment = self.create(CommentItem, UML.Comment)
+        line = self.create(CommentLineItem)
 
         self.connect(line, line.head, comment)
         # connected, but no annotated element yet
@@ -30,8 +33,8 @@ class CommentLineTestCase(TestCase):
     def test_commentline_same_comment_glue(self):
         """Test comment line item gluing to already connected comment item."""
 
-        comment = self.create(items.CommentItem, UML.Comment)
-        line = self.create(items.CommentLineItem)
+        comment = self.create(CommentItem, UML.Comment)
+        line = self.create(CommentLineItem)
 
         self.connect(line, line.head, comment)
         glued = self.allow(line, line.tail, comment)
@@ -40,9 +43,9 @@ class CommentLineTestCase(TestCase):
     def test_commentline_element_connect(self):
         """Test comment line connecting to comment and actor items.
         """
-        comment = self.create(items.CommentItem, UML.Comment)
-        line = self.create(items.CommentLineItem)
-        ac = self.create(items.ActorItem, UML.Actor)
+        comment = self.create(CommentItem, UML.Comment)
+        line = self.create(CommentLineItem)
+        ac = self.create(ActorItem, UML.Actor)
 
         self.connect(line, line.head, comment)
         self.connect(line, line.tail, ac)
@@ -53,9 +56,9 @@ class CommentLineTestCase(TestCase):
     def test_commentline_element_connect(self):
         """Test comment line connecting to comment and actor items.
         """
-        comment = self.create(items.CommentItem, UML.Comment)
-        line = self.create(items.CommentLineItem)
-        ac = self.create(items.ActorItem, UML.Actor)
+        comment = self.create(CommentItem, UML.Comment)
+        line = self.create(CommentLineItem)
+        ac = self.create(ActorItem, UML.Actor)
 
         self.connect(line, line.head, comment)
         self.connect(line, line.tail, ac)
@@ -66,9 +69,9 @@ class CommentLineTestCase(TestCase):
     def test_commentline_element_reconnect(self):
         """Test comment line connecting to comment and actor items.
         """
-        comment = self.create(items.CommentItem, UML.Comment)
-        line = self.create(items.CommentLineItem)
-        ac = self.create(items.ActorItem, UML.Actor)
+        comment = self.create(CommentItem, UML.Comment)
+        line = self.create(CommentLineItem)
+        ac = self.create(ActorItem, UML.Actor)
 
         self.connect(line, line.head, comment)
         self.connect(line, line.tail, ac)
@@ -76,7 +79,7 @@ class CommentLineTestCase(TestCase):
         self.assertEqual(1, len(comment.subject.annotatedElement))
         self.assertTrue(ac.subject in comment.subject.annotatedElement)
 
-        ac2 = self.create(items.ActorItem, UML.Actor)
+        ac2 = self.create(ActorItem, UML.Actor)
         # ac.canvas.disconnect_item(line, line.tail)
         self.disconnect(line, line.tail)
         self.connect(line, line.tail, ac2)
@@ -87,9 +90,9 @@ class CommentLineTestCase(TestCase):
     def test_commentline_element_disconnect(self):
         """Test comment line connecting to comment and disconnecting actor item.
         """
-        comment = self.create(items.CommentItem, UML.Comment)
-        line = self.create(items.CommentLineItem)
-        ac = self.create(items.ActorItem, UML.Actor)
+        comment = self.create(CommentItem, UML.Comment)
+        line = self.create(CommentLineItem)
+        ac = self.create(ActorItem, UML.Actor)
 
         self.connect(line, line.head, comment)
         self.connect(line, line.tail, ac)
@@ -102,9 +105,9 @@ class CommentLineTestCase(TestCase):
     def test_commentline_unlink(self):
         """Test comment line unlinking.
         """
-        clazz = self.create(items.ClassItem, UML.Class)
-        comment = self.create(items.CommentItem, UML.Comment)
-        line = self.create(items.CommentLineItem)
+        clazz = self.create(ClassItem, UML.Class)
+        comment = self.create(CommentItem, UML.Comment)
+        line = self.create(CommentLineItem)
 
         self.connect(line, line.head, comment)
         self.connect(line, line.tail, clazz)
@@ -131,9 +134,9 @@ class CommentLineTestCase(TestCase):
     def test_commentline_element_unlink(self):
         """Test comment line unlinking using a class item.
         """
-        clazz = self.create(items.ClassItem, UML.Class)
-        comment = self.create(items.CommentItem, UML.Comment)
-        line = self.create(items.CommentLineItem)
+        clazz = self.create(ClassItem, UML.Class)
+        comment = self.create(CommentItem, UML.Comment)
+        line = self.create(CommentLineItem)
 
         self.connect(line, line.head, comment)
         self.connect(line, line.tail, clazz)
@@ -159,9 +162,9 @@ class CommentLineTestCase(TestCase):
 
         Demonstrates defect #103.
         """
-        clazz1 = self.create(items.ClassItem, UML.Class)
-        clazz2 = self.create(items.ClassItem, UML.Class)
-        gen = self.create(items.GeneralizationItem)
+        clazz1 = self.create(ClassItem, UML.Class)
+        clazz2 = self.create(ClassItem, UML.Class)
+        gen = self.create(GeneralizationItem)
 
         self.connect(gen, gen.head, clazz1)
         self.connect(gen, gen.tail, clazz2)
@@ -169,8 +172,8 @@ class CommentLineTestCase(TestCase):
         assert gen.subject
 
         # now, connect comment to a generalization (relationship)
-        comment = self.create(items.CommentItem, UML.Comment)
-        line = self.create(items.CommentLineItem)
+        comment = self.create(CommentItem, UML.Comment)
+        line = self.create(CommentLineItem)
         self.connect(line, line.head, comment)
         self.connect(line, line.tail, gen)
 
@@ -186,11 +189,11 @@ class CommentLineTestCase(TestCase):
         """
         It is not allowed to create two commentlines between the same elements.
         """
-        clazz = self.create(items.ClassItem, UML.Class)
+        clazz = self.create(ClassItem, UML.Class)
 
         # now, connect comment to a generalization (relationship)
-        comment = self.create(items.CommentItem, UML.Comment)
-        line1 = self.create(items.CommentLineItem)
+        comment = self.create(CommentItem, UML.Comment)
+        line1 = self.create(CommentLineItem)
         self.connect(line1, line1.head, comment)
         self.connect(line1, line1.tail, clazz)
 
@@ -199,7 +202,7 @@ class CommentLineTestCase(TestCase):
 
         # Now add another line
 
-        line2 = self.create(items.CommentLineItem)
+        line2 = self.create(CommentLineItem)
         self.connect(line2, line2.head, comment)
 
         self.assertFalse(self.allow(line2, line2.tail, clazz))
