+++ gaphor/diagram/interfaces.py
@@ -22,18 +22,3 @@ def Editor(obj):
 class Group:
     def __init__(self, parent, item):
         pass
-
-
-@multidispatch(object, object)
-class IConnect:
-    """
-    This function is used by the HandleTool to allow connecting
-    lines to element items. For each specific case (Element, Line) an
-    adapter could be written.
-    """
-
-    def __init__(self, item, line_item):
-        pass
-
-    def allow(self, handle, port):
-        return False
