+++ gaphor/tests/testcase.py
@@ -13,7 +13,7 @@ from gaphas.aspect import ConnectionSink, Connector
 
 from gaphor import UML
 from gaphor.application import Application
-from gaphor.diagram.interfaces import IConnect
+from gaphor.diagram.connectors import IConnect
 from gaphor.diagram.interfaces import Group
 
 # For DiagramItemConnector aspect:
