+++ gaphor/ui/tests/test_handletool.py
@@ -13,7 +13,7 @@ from gaphor.core import inject
 from gaphor.diagram.usecases.actor import ActorItem
 from gaphor.diagram.general.comment import CommentItem
 from gaphor.diagram.general.commentline import CommentLineItem
-from gaphor.diagram.interfaces import IConnect
+from gaphor.diagram.connectors import IConnect
 from gaphor.ui.diagramtools import ConnectHandleTool, DiagramItemConnector
 from gaphor.ui.event import DiagramShow
 from gaphor.ui.abc import UIComponent
